requirements.txt
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.12.1
pytz==2020.10
gunicorn==20.0.4
whitenoise==5.2.0